Rename and move `udpateledgers` to bookie group
