+++ tools/ledger/src/main/java/org/apache/bookkeeper/tools/cli/commands/LedgerCommandGroup.java
@@ -23,7 +23,6 @@ import static org.apache.bookkeeper.tools.common.BKCommandCategories.CATEGORY_LE
 import org.apache.bookkeeper.tools.cli.BKCtl;
 import org.apache.bookkeeper.tools.cli.commands.client.DeleteLedgerCommand;
 import org.apache.bookkeeper.tools.cli.commands.client.SimpleTestCommand;
-import org.apache.bookkeeper.tools.cli.commands.client.UpdateLedgersCommand;
 import org.apache.bookkeeper.tools.common.BKFlags;
 import org.apache.bookkeeper.tools.framework.CliCommandGroup;
 import org.apache.bookkeeper.tools.framework.CliSpec;
@@ -43,7 +42,6 @@ public class LedgerCommandGroup extends CliCommandGroup<BKFlags> {
         .withCategory(CATEGORY_LEDGER_SERVICE)
         .addCommand(new SimpleTestCommand())
         .addCommand(new DeleteLedgerCommand())
-        .addCommand(new UpdateLedgersCommand())
         .build();
 
     public LedgerCommandGroup() {
